+++ test/dummy/app/models/user.rb
@@ -2,5 +2,7 @@ class User < ApplicationRecord
   validates :name, presence: true
   validates :email, presence: true, uniqueness: {case_sensitive: false}
 
+  normalizes :email, with: -> { _1.strip.downcase }
+
   authenticates
 end