+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## HEAD
 
+* Normalise identifier value for lookup.
+
 
 ## 2.2.4 (25 June 2024)